# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Herramientas/ParaOferta": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetHerramientasParaOferta",
        "parameters": [
          {
            "name": "filtroFabricante",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroPrecio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramientas/Para-Alquiler": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetHerramientasParaAlquiler",
        "parameters": [
          {
            "name": "filtroNombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroMaterial",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaAlquilarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaAlquilarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaAlquilarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramientas/Para-Compra": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetHerramientasParaCompra",
        "parameters": [
          {
            "name": "filtroMaterial",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroPrecio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramientas/Para-Reparación": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetHerramientasParaReparación",
        "parameters": [
          {
            "name": "filtroNombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroTiempoReparacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaReparaciónDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaReparaciónDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientasParaReparaciónDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HerramientasParaAlquilarDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "HerramientasParaComprarDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "HerramientasParaOfertarDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "HerramientasParaReparaciónDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "tiempoReparación": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}